add unit test for updateDependencyVersion
